fixing realtime repository
