+++ app/build.gradle
@@ -64,6 +64,7 @@ dependencies {
     kapt 'com.google.dagger:hilt-android-compiler:2.45'
 
     implementation platform('com.google.firebase:firebase-bom:31.2.3')
+    implementation 'com.google.firebase:firebase-crashlytics-ktx'
     implementation 'com.google.firebase:firebase-analytics-ktx'
     implementation 'com.google.firebase:firebase-database-ktx'
     implementation 'com.google.firebase:firebase-auth-ktx'
